fix: OutputView 사용자 안내 메세지 출력하는 메서드 로직 수정 - feat: 콤마를 포함해 사용자 이름을 담은 String을 반환하는 메서드 추가 feat: 게임 결과 정보 담고 있는 ResultStatistics model class 추가
